Take a hint or two from Clippy
`Empty` variants had all the same prefix. Move it to the enum name,
now `EmptyIs`. Reads the same, and the variants are more
distinguishable.

All integer and string decoding functions apparently don't need to
take a vector, but are fine with a slice. This elides creating a new
Vec object.

Remove needless `return`s.

Single-case matches are better represented by an `if let` construct.

A `match` with a pattern evaluating to true and the rest to false is a
case for the `matches!()` macro.