Move dynamic field pos into new FieldOffset enum
Previously the filename field in GlobalVariables was retrieved through
a function because it's at a variable offset, depending on the fields
before it.

Now only the _offset_ is retrieved through a function, and the field
is retrieved normally. This makes more sense because now in the
ResourceSpec for GlobalVariables both the offset has meaningful
content and the ftype is actually the type of the field, instead of
that being an implementation detail of a function way below.