Add APE file finding feature
Scan all .ape files in the given --winamp-dir for possible ID strings.

When all resources are collected (and -a/--find-apes is given) search
the list of APE file strings collected earlier and if a match was
found replace the `Ape` resource with an `ApeFile` resource,
containing the filename.

`--winamp-dir` does not do "Plugins/avs" subdir resolving yet, you
have to pass "path/to/Winamp/Plugins/avs/".