Move ouput printing out of lib.rs
The lib now returns the gathered data and does not print to stdout.

Consequently, move the yaml-escape function to main as well, because
it's not used in lib anymore and doesn't need to be public.

Make `ResourceType`, `Resource` and `Arguments` as well as their
fields public for lib, since these are now the input and output data
types.