Split path string decoding on Windows & Linux
While Linux file paths are generally UTF-8, they don't have to be.
Consequently on Linux you may get UTF-8 decoding errors when
converting a path to a string. In that case the code falls back to
assuming (blindly) that the path is windows-1252-encoded (a.k.a.
latin1), which works for most cases.

Before this the code would not compile on Windows however because the
Windows version of OsStrExt, the platform-specific extension of a rust
path string, has no such concept of decoding to UTF-8. What exactly
the Windows way of dealing with legacy file path encodings is I still
have to find out, but for now, just assume that the paths are valid
UTF-8 strings.